requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
pytorch-lightning>=2.0
statsmodels>=0.14
yfinance>=0.2
fredapi>=0.5
requests>=2.28
fastapi>=0.95
uvicorn>=0.22
websockets>=10.4
streamlit>=1.20
dash>=2.9
plotly>=5.14
matplotlib>=3.7
seaborn>=0.12
shap>=0.41
hmmlearn>=0.3
arch>=5.6
scipy>=1.10
pytest>=7.2
python-dotenv>=1.0
pydantic>=1.10
bayesian-torch>=0.2 